python-backend/requirements.txt
Pillow==10.4.0
imagehash==4.3.1
opencv-contrib-python==4.10.0.84
numpy
onnxruntime
mediapipe
rawpy
